Add eval results & permalinks (#62)
This PR's primary goal is to add links to spans and experiments. To do this, we need to login and get the current org-name (which is part of the URL), and pass more data into evals (orgName, etc).

This PR implements auth & auth caching to support this. 

This is a backwards incompatible change. It changes the `Eval.Run()` function to
return `(Result, error)` instead of `error` and `eval.New()` takes a `eval.Key`
instead of an experiment ID.

The full changes and migration guide are described here

https://gist.github.com/clutchski/209af31eb8899334e09b9738204e9f9d